configure plugins for global css
