+++ Barsamyan-D/str-strip-plugin-david/structure_strip.py
@@ -1,42 +1,15 @@
 # structure_strip.py
-import marimo
-import anywidget
-import traitlets
 
-__generated_with = "0.10.12"
+import marimo
+
+__generated_with = "0.11.7"
 app = marimo.App(width="medium")
 
+
 @app.cell
-def __():
-    import anywidget
-    import traitlets
-    import pathlib
+def _():
+    from StructureStripWidget import StructureStripWidget
 
-    assets_dir = pathlib.Path(__file__).parent / "assets"
-
-    class StructureStripWidget(anywidget.AnyWidget):
-        _esm = "str.js"
-        _css = "str.css"
-            
-        sections = traitlets.List().tag(sync=True)
-        user_inputs = traitlets.Dict().tag(sync=True)
-        title = traitlets.Unicode().tag(sync=True)
-        description = traitlets.Unicode().tag(sync=True)
-        image_path = traitlets.Unicode().tag(sync=True)
-
-        def __init__(self, **kwargs):
-            super().__init__(**kwargs)
-            image_path = assets_dir / "london.jpg"
-            self.image_path = self._file_to_data_url(image_path)
-            #str(assets_dir / "london.jpg")
-            #self.set("image_path", self.image_path)
-        
-        def _file_to_data_url(self, file_path):
-            import base64
-            mime_type = "image/jpeg"
-            data = base64.b64encode(file_path.read_bytes()).decode("utf-8")
-            return f"data:{mime_type};base64,{data}"
-        
     structure_strip = StructureStripWidget(
         title="London Docklands Evaluation",
         description="Make yourself familiar with the Docklands in London that underwent major changes. To what extend was the Docklands Regeneration successful? Your evaluation of the successes and the failures each should be roughly three times the size of your introduction and your conclusion.",
@@ -73,8 +46,8 @@ def __():
     )
 
     structure_strip
+    return StructureStripWidget, structure_strip
 
-    return structure_strip,
 
 if __name__ == "__main__":
     app.run()
