+++ cassandratin13/sort_paragraphs_plugin/sort_paragraphs.py
@@ -1,26 +1,19 @@
 import marimo
 
-__generated_with = "0.10.9"
+__generated_with = "0.11.7"
 app = marimo.App(width="medium")
 
 
 @app.cell
 def _():
     import marimo as mo
-    import anywidget
-    import traitlets
-
-    class SortTheParagraphs(anywidget.AnyWidget):
-        _esm = "stp.js"
-        _css = "stp.css"
-        question = traitlets.Unicode(default_value="Sort the texts")
-        sorted_texts = traitlets.List(default_value=["Text 1", "Text 2", "Text 3", "Text 4"]).tag(sync=True)
+    from SortTheParagraphs import SortTheParagraphs
 
     question = "Order the steps for problem solving."
     texts = ["Understand the problem", "Make a plan", "Carry out the plan", "Look back and reflect"]
 
     SortTheParagraphs(question=question, sorted_texts=texts)
-    return SortTheParagraphs, anywidget, mo, question, texts, traitlets
+    return SortTheParagraphs, mo, question, texts
 
 
 @app.cell
